separate soliidty tests in CI
